+++ src/app/data-manager/page.tsx
@@ -1,0 +1,13 @@
+import { DataManagerToolbar } from '@/components/layout';
+
+export default function DataManagerPage() {
+  return (
+    <div className="flex flex-column min-h-screen">
+      <DataManagerToolbar />
+      <div className="flex-1 p-4">
+        {/* Main content area - to be developed */}
+      </div>
+    </div>
+  );
+}
+